fix c1-c3 comments

+++ src/AlgoTradeForge.Domain/Live/LiveSessionConfig.cs
@@ -8,7 +8,7 @@ public sealed record LiveSessionConfig
     public required IInt64BarStrategy Strategy { get; init; }
     public required IList<DataSubscription> Subscriptions { get; init; }
     public required Asset PrimaryAsset { get; init; }
-    public required decimal InitialCash { get; init; }
+    public required long InitialCash { get; init; }
     public long CommissionPerTrade { get; init; }
     public LiveEventRouting Routing { get; init; } = LiveEventRouting.OnBarComplete | LiveEventRouting.OnTrade;
     public required string AccountName { get; init; }
